requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Array operations
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0

# Image loading
Pillow>=10.0.0

# Metrics & data splitting
scikit-learn>=1.3.0
